backend/requirements.txt
# backend/requirements.txt

fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
requests==2.31.0
pydantic==2.6.1
psycopg2-binary==2.9.9
python-dotenv==0.21.1